Change abstract class name.

+++ src/shared_modules/dbsync/src/dbsync_implementation.h
@@ -57,7 +57,7 @@ namespace DbSync
 
             void setMaxRows(const DBSYNC_HANDLE handle,
                             const std::string& table,
-                            const int64_t maxRows);
+                            const long long maxRows);
 
             TXN_HANDLE createTransaction(const DBSYNC_HANDLE    handle,
                                          const nlohmann::json&  json);
